Backend/tests/unit/website/test_custom.py
# from django.test import SimpleTestCase
# from datetime import datetime
# from django.template import Context, Template
# from django.utils import timezone

# class CustomTemplateFiltersTestCase(SimpleTestCase):
#     def test_is_alternate_weekday(self):
#         # Test for the 'is_alternate_weekday' filter
#         template = Template("{% load your_actual_template_tags_library %}{{ '2023-10-14'|is_alternate_weekday:2 }}")
#         rendered = template.render(Context({}))
#         self.assertEqual(rendered, "Your expected output here")

#     def test_get_occurrence_value(self):
#         # Test for the 'get_occurrence_value' filter
#         task_media_dict = {
#             1: {
#                 1: {1: True}
#             }
#         }
#         template = Template("{% load your_actual_template_tags_library %}{{ '111'|get_occurrence_value:task_media_dict }}")
#         rendered = template.render(Context({'task_media_dict': task_media_dict}))
#         self.assertEqual(rendered, "Y")

#     def test_concat(self):
#         # Test for the 'concat' filter
#         template = Template("{% load your_actual_template_tags_library %}{{ 'abc'|concat:'def' }}")
#         rendered = template.render(Context({}))
#         self.assertEqual(rendered, "abcdef")

#     # Add more test cases for other filters similarly

